Move ert.shared.services to ert.services
