+++ .github/workflows/style.yml
@@ -49,7 +49,7 @@ jobs:
     - name: Run pylint
       run: |
         pip install .  # We need the dependencies of ERT to avoid import-error
-        pylint src/ert --ignore-paths=src/ert/_c_wrappers,src/ert/gui,src/ert/logging,src/ert/analysis/_es_update.py,src/ert/data,src/ert/job_runner,src/ert/shared
+        pylint src/ert --ignore-paths=src/ert/_c_wrappers,src/ert/gui,src/ert/logging,src/ert/analysis/_es_update.py,src/ert/data,src/ert/job_runner,src/ert/shared,src/ert/services
         pylint --disable=all \
           --enable="\
             access-member-before-definition, \
