+++ .mypy-strict.ini
@@ -6,10 +6,14 @@ exclude = (?x)(
     | src/ert/ert3/examples
     | src/ert/shared
     | src/ert/dark_storage
+    | src/ert/services
     | src/ert/_c_wrappers
   )
 
 [mypy-ert.shared.*]
+ignore_errors = True
+
+[mypy-ert.services.*]
 ignore_errors = True
 
 [mypy-ert.gui.*]
