+++ src/ert/services/_base_service.py
