+++ src/ert/storage/_storage.py
@@ -11,7 +11,7 @@ import pandas as pd
 import ert.data
 import ert.exceptions
 from ert.async_utils import get_event_loop
-from ert.shared.services import Storage
+from ert.services import Storage
 
 logger = logging.getLogger(__name__)
 read_csv = partial(pd.read_csv, index_col=0, float_precision="round_trip")
