+++ tests/ert_tests/services/test_storage_service.py
@@ -4,8 +4,8 @@ import os
 import pytest
 import requests
 
+from ert.services import Storage, _storage_main
 from ert.shared import port_handler
-from ert.shared.services import Storage, _storage_main
 
 
 @pytest.mark.skip
